+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - ✨ Add a custom callout block to the editor #892
 - 🚩(frontend) version MIT only #911
+- ✨(backend) integrate maleware_detection from django-lasuite #936
 
 ## Changed